ai_app/requirements.txt
streamlit
pandas
bcrypt
faiss-cpu
sentence-transformers
openai
